+++ Pipfile
@@ -32,6 +32,6 @@ python_version = "3.10"
 allow_prereleases = true
 
 [scripts]
-server = "hypercorn app.main:app --reload --bind 0.0.0.0:8080"
+server = "hypercorn apex_res.main:app --reload --bind 0.0.0.0:8080"
 apex = "python main.py"
 test = "pytest -vv"
